Added command permissions

+++ commands/clashroles.js
@@ -31,53 +31,62 @@ module.exports = {
     async execute(interaction) {
 
         if (interaction.options.getSubcommand() == 'edit') {
-            // TODO: only certain role members should be allowed to use this command (permissions!)
-
             let reply = 'Unknown error.';
-            const townhall = interaction.options.get('townhall').value;
-            const role = interaction.options.get('role').value;
 
-            const throle = await db_storage_handler.getTownhallRole(townhall);
+            // only certain members should be allowed to use this command (permissions!)
+            if (interaction.member.permissions.has(Permissions.FLAGS.MANAGE_ROLES)) {
+                const townhall = interaction.options.get('townhall').value;
+                const role = interaction.options.get('role').value;
 
-            if (throle) {
-                await db_storage_handler.editTownhallRole(townhall, role)
-                    .then(() => {
-                        reply = roleMention(role) + ' is now the new role for townhall ' + townhall + '.';
-                    }).catch((error) => {
-                        console.log(error);
-                        reply = 'Error updating the database: ' + error.message;
-                    });
+                const throle = await db_storage_handler.getTownhallRole(townhall);
+
+                if (throle) {
+                    await db_storage_handler.editTownhallRole(townhall, role)
+                        .then(() => {
+                            reply = roleMention(role) + ' is now the new role for townhall ' + townhall + '.';
+                        }).catch((error) => {
+                            console.log(error);
+                            reply = 'Error updating the database: ' + error.message;
+                        });
+                } else {
+                    await db_storage_handler.addTownhallRole(townhall, role)
+                        .then((entry) => {
+                            reply = roleMention(entry.role_id) + ' for townhall ' + entry.townhall + ' added.';
+                        }).catch((error) => {
+                            console.log(error);
+                            reply = 'Error inserting into database: ' + error.message;
+                        });
+                }
             } else {
-                await db_storage_handler.addTownhallRole(townhall, role)
-                    .then((entry) => {
-                        reply = roleMention(entry.role_id) + ' for townhall ' + entry.townhall + ' added.';
-                    }).catch((error) => {
-                        console.log(error);
-                        reply = 'Error inserting into database: ' + error.message;
-                    });
+                reply = 'You are not allowed to do this.';
             }
 
             await interaction.reply(reply);
 
         } else if (interaction.options.getSubcommand() == 'delete') {
-            // TODO: only certain role members should be allowed to use this command (permissions!)
-
             let reply = 'Unknown error.';
-            const townhall = interaction.options.get('townhall').value;
 
-            await db_storage_handler.deleteTownhallRole(townhall)
-                .then((count) => {
+            // only certain members should be allowed to use this command (permissions!)
+            if (interaction.member.permissions.has(Permissions.FLAGS.MANAGE_ROLES)) {
+                const townhall = interaction.options.get('townhall').value;
 
-                    if (count == 0) {
-                        reply = 'A role for townhall ' + townhall + ' doesn\'t exist yet.';
-                    } else {
-                        reply = 'Successfully deleted role entry for townhall ' + townhall + '.';
-                    }
+                await db_storage_handler.deleteTownhallRole(townhall)
+                    .then((count) => {
 
-                }).catch((error) => {
-                    console.log(error);
-                    reply = 'Error deleting stuff from the database: ' + error.message;
-                });
+                        if (count == 0) {
+                            reply = 'A role for townhall ' + townhall + ' doesn\'t exist yet.';
+                        } else {
+                            reply = 'Successfully deleted role entry for townhall ' + townhall + '.';
+                        }
+
+                    }).catch((error) => {
+                        console.log(error);
+                        reply = 'Error deleting stuff from the database: ' + error.message;
+                    });
+
+            } else {
+                reply = 'You are not allowed to do this.'
+            }
 
             await interaction.reply(reply);
